Move chat views to Views/Chat subfolder and pin package versions
- Move ChatOverlayView, ChatBubbleTemplateSelector, MarkdownConverter to Views/Chat/
- Update namespaces to LocalChatClientWithAgents.Views.Chat
- Pin Microsoft.Agents.AI packages to 1.0.0-rc3 (Hosting: 1.0.0-preview.260304.1)
- Pin Microsoft.Extensions.AI to 10.3.0
- Pin Microsoft.Extensions.* to 10.0.0
- Fix Microsoft.Maui.Essentials.AI version glob to 10.0.*-*
