+++ 10.0/AI/LocalChatClientWithAgents/src/Pages/LandmarksPage.xaml.cs
@@ -1,6 +1,6 @@
 using LocalChatClientWithAgents.Models;
 using LocalChatClientWithAgents.ViewModels;
-using LocalChatClientWithAgents.Views;
+using LocalChatClientWithAgents.Views.Chat;
 
 namespace LocalChatClientWithAgents.Pages;
