+++ 10.0/AI/LocalChatClientWithAgents/src/Views/Chat/ChatOverlayView.xaml
@@ -2,9 +2,9 @@
 <ContentView
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:views="clr-namespace:LocalChatClientWithAgents.Views"
+    xmlns:views="clr-namespace:LocalChatClientWithAgents.Views.Chat"
     xmlns:models="clr-namespace:LocalChatClientWithAgents.ViewModels"
-    x:Class="LocalChatClientWithAgents.Views.ChatOverlayView">
+    x:Class="LocalChatClientWithAgents.Views.Chat.ChatOverlayView">
 
     <ContentView.Resources>
         <views:MarkdownConverter x:Key="MarkdownConverter" />
